socket modifications handling prod/dev connections

+++ vscheduler/socket/vis_client_login.py
@@ -20,18 +20,19 @@ def main():
             client.connect(ADDR)
         except socket.error as e:
             print (f"Caught exception socket.error: {e}")
-            mailFunction(f"socket error - {host}", f"error connecting managment socket server ({ADDR[0][0]}:{ADDR[0][1]}) from {host} at LOGIN attempt for {user}\n{str(e)}")
+            mailFunction(f"socket error - {host}", f"error connecting managment socket server ({ADDR[0]}:{ADDR[1]}) from {host} at LOGIN attempt for {user}\n{str(e)}")
             os.system(f'pkill -KILL -u {user}')
 
-        print(f"[CONNECTED] VIS CLIENT TO MGMT SERVER AT {ADDR[0][0]}:{ADDR[0][1]}")
+        print(f"[CONNECTED] VIS CLIENT TO MGMT SERVER AT {ADDR[0]}:{ADDR[0]}")
         
         while connected:
             msg = [host, user, operating_system]
             client.send((msg[0] + "," + msg[1] + "," + msg[2]).encode(FORMAT))
             msg = client.recv(SIZE)
             print(f"[MGMT SERVER] sent: {msg}")
-            os.system(f'echo "/usr/bin/python3 /etc/profile.d/vis_client_logout.py" | at now +{msg[1]} hour')
-            os.system(f'echo "pkill -9 -u $USER" | at now +{msg[1] + 1} minute')
+            if msg[2] == "up":
+                os.system(f'echo "/usr/bin/python3 /etc/profile.d/vis_client_logout.py" | at now +{msg[1]} hour')
+                os.system(f'echo "pkill -9 -u $USER" | at now +{msg[1] + 1} minute')
             connected = False
         if msg[2] != "dev":
             break
